re organize the code a bit
